chore: move and split file_parser.py
